use uint64 type for stakedtokens as int32 is too small

+++ internal/packages/consensus/uptime/types/types_common.go
@@ -26,7 +26,7 @@ type ValidatorUptimeStatus struct {
 	ValidatorConsensusAddress string  `json:"validator_consensus_addreess"`
 	MissedBlockCounter        float64 `json:"missed_block_counter"`
 	IsTomstoned               float64
-	StakedTokens              int
+	StakedTokens              uint64
 	CommissionRate            float64
 	// Only Consumer Chain
 	ConsumerConsensusAddress string `json:"consumer_consensus_address"`
